Move `Counter` out into its own source file
